LPD-15752 portal-search-tuning-rankings-web: Add check to see if table exists
https://liferay.atlassian.net/browse/LPD-15752

+++ modules/dxp/apps/portal-search-tuning/portal-search-tuning-rankings-web/src/main/java/com/liferay/portal/search/tuning/rankings/web/internal/upgrade/v2_0_0/RenameRankingUpgradeProcess.java
@@ -33,7 +33,7 @@ public class RenameRankingUpgradeProcess extends UpgradeProcess {
 				"com.liferay.portal.search.tuning.rankings.web.internal." +
 					"index.Ranking");
 
-			if (className != null) {
+			if ((className != null) && hasTable("JSONStorageEntry")) {
 				_classNameLocalService.deleteClassName(className);
 
 				ClassName newClassName = _classNameLocalService.getClassName(
